INSPECTION ALIMENTAIRE/requirements.txt
aniso8601==8.0.0
api==0.0.7
APScheduler==2.1.2
beautifulsoup4==4.8.2
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
dicttoxml==1.7.4
Flask==1.1.1
Flask-Jsonpify==1.5.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.1
flask-whooshalchemy3==0.1.2
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
json2xml==3.3.2
lxml==4.5.0
MarkupSafe==1.1.1
nose==1.3.7
oauthlib==3.1.0
pbr==3.1.1
PySocks==1.7.1
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.11.0
soupsieve==2.0
SQLAlchemy==1.1.13
tweepy==3.8.0
tzlocal==2.0.0
urllib3==1.25.8
Werkzeug==1.0.0
Whoosh==2.7.4
xmltodict==0.11.0
